requirements.txt
Flask
twitter_scraper
httplib2
request